add search-storage, image-storage
